Fixed the unreliable two burn orbit raise test case. (#775)
* Gave the finite burn orbit raise test cases a common function for building the test cases, and used the IPOPT settings that should provide more robust convergence.

* More cleanup of the finite burn orbit raise.

* Even at zero iterations IPOPT was changing design variables. Changed load_case to allow either a Case or a Problem to be compared.

* a few more tweaks to the case comparison

* few more tweaks

* Fixed MPI issue, dm.run_problem simulate option is causing hangs in testflo even though it works under normal MPI, logging issue.

* tolerance tweak for the radau restart test
