+++ dymos/examples/finite_burn_orbit_raise/test/test_ex_two_burn_orbit_raise_mpi.py
@@ -12,36 +12,21 @@ from dymos.examples.finite_burn_orbit_raise.finite_burn_orbit_raise_problem impo
 
 
 @require_pyoptsparse(optimizer='IPOPT')
+@unittest.skipUnless(MPI, "MPI is required.")
 @use_tempdirs
-class TestExampleTwoBurnOrbitRaise(unittest.TestCase):
+class TestExampleTwoBurnOrbitRaiseMPI(unittest.TestCase):
+    N_PROCS = 3
 
     def test_ex_two_burn_orbit_raise(self):
         optimizer = 'IPOPT'
 
         p = two_burn_orbit_raise_problem(transcription='gauss-lobatto', transcription_order=3,
-                                         compressed=False, optimizer=optimizer,
+                                         compressed=False, optimizer=optimizer, simulate=False,
                                          show_output=False)
 
         if p.model.traj.phases.burn2 in p.model.traj.phases._subsystems_myproc:
             assert_near_equal(p.get_val('traj.burn2.states:deltav')[-1], 0.3995,
                               tolerance=2.0E-3)
-
-
-# This test is separate because connected phases aren't directly parallelizable.
-@require_pyoptsparse(optimizer='IPOPT')
-@use_tempdirs
-class TestExampleTwoBurnOrbitRaiseConnected(unittest.TestCase):
-
-    def test_ex_two_burn_orbit_raise_connected(self):
-        optimizer = 'IPOPT'
-
-        p = two_burn_orbit_raise_problem(transcription='gauss-lobatto', transcription_order=3,
-                                         compressed=False, optimizer=optimizer,
-                                         show_output=False, connected=True)
-
-        if p.model.traj.phases.burn2 in p.model.traj.phases._subsystems_myproc:
-            assert_near_equal(p.get_val('traj.burn2.states:deltav')[0], 0.3995,
-                              tolerance=4.0E-3)
 
 
 if __name__ == '__main__':  # pragma: no cover
